proyectoSomnolenciaFinal/requirements.txt
absl-py==2.1.0
attrs==24.3.0
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.1
chardet==5.2.0
contourpy==1.3.1
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.9
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.4
jax==0.5.0
jaxlib==0.5.0
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
mediapipe==0.10.20
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
protobuf==4.25.5
psutil==6.1.1
psycopg2==2.9.10
pycparser==2.22
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
reportlab==4.2.5
scipy==1.15.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
svglib==1.5.1
tinycss2==1.4.0
virtualenv==20.28.0
webencodings==0.5.1
